Add controller tests

+++ product-catalog-service/src/main/java/com/rainbowforest/productcatalogservice/controller/AdminProductController.java
@@ -1,15 +1,13 @@
 package com.rainbowforest.productcatalogservice.controller;
 
-import com.rainbowforest.productcatalogservice.model.Product;
+import com.rainbowforest.productcatalogservice.entity.Product;
+import com.rainbowforest.productcatalogservice.http.header.HeaderGenerator;
 import com.rainbowforest.productcatalogservice.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
-import java.net.URI;
-import java.net.URISyntaxException;
 
 @RestController
 @RequestMapping("/admin")
@@ -17,18 +15,47 @@ public class AdminProductController {
 
     @Autowired
     private ProductService productService;
+    
+    @Autowired
+    private HeaderGenerator headerGenerator;
 
     @PostMapping(value = "/products")
-    private ResponseEntity<Product> addProduct(@RequestBody Product product, HttpServletRequest request) throws URISyntaxException {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setLocation(new URI(request.getRequestURI() + "/" + product.getId()));
-        productService.addProduct(product);
-        return new ResponseEntity<Product>(product, headers, HttpStatus.CREATED);
+    private ResponseEntity<Product> addProduct(@RequestBody Product product, HttpServletRequest request){
+    	if(product != null) {
+    		try {
+    			productService.addProduct(product);
+    	        return new ResponseEntity<Product>(
+    	        		product,
+    	        		headerGenerator.getHeadersForSuccessPostMethod(request, product.getId()),
+    	        		HttpStatus.CREATED);
+    		}catch (Exception e) {
+				e.printStackTrace();
+				return new ResponseEntity<Product>(
+						headerGenerator.getHeadersForError(),
+						HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+    	}
+    	return new ResponseEntity<Product>(
+    			headerGenerator.getHeadersForError(),
+    			HttpStatus.BAD_REQUEST);       
     }
 
     @DeleteMapping(value = "/products/{id}")
     private ResponseEntity<Void> deleteProduct(@PathVariable("id") Long id){
-        productService.deleteProduct(id);
-        return new ResponseEntity<Void>(HttpStatus.OK);
+    	Product product = productService.getProductById(id);
+    	if(product != null) {
+    		try {
+    			productService.deleteProduct(id);
+    	        return new ResponseEntity<Void>(
+    	        		headerGenerator.getHeadersForSuccessGetMethod(),
+    	        		HttpStatus.OK);
+    		}catch (Exception e) {
+				e.printStackTrace();
+    	        return new ResponseEntity<Void>(
+    	        		headerGenerator.getHeadersForError(),
+    	        		HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+    	}
+    	return new ResponseEntity<Void>(headerGenerator.getHeadersForError(), HttpStatus.NOT_FOUND);      
     }
 }
